payments/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "payments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "kafka-python>=2.2.15",
    "llama-index-workflows>=2.6.0",
    "pydantic>=2.11.10",
    "sqlalchemy>=2.0.43",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/payments"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[project.scripts]
process-payments = "payments.workflow.main:main"